Implement item stock and rating update endpoints

+++ microservices/item/src/main/java/beady/shop/item/dto/ReduceItemDTO.java
@@ -13,5 +13,5 @@ public class ReduceItemDTO {
 
 	@NotNull
 	@Min(0)
-	private Long reduceBy;
+	private Integer reduceBy;
 }
